Separate component for Delete Account.
